set: only fetch the device values if we're not supplying all of them
On some devices, get_feature_report will fail with errors but,
presumably, it's still possible to write a feature report to the device
and have it take effect.

So let's revamp the code that we only fill in the values if the user
didn't give us all of those. This makes it possible to e.g. run
   hid-feature set --report-id=4 ab 12 cd 34
and have it succeed, even where GetFeature would otherwise fail.