Allow skipping report IDs in list
Some devices are happy to have all the feature reports listed except for
one or more specific ones. Let's allow skipping those.

And to make that possible let's print the error of the one that fails
before we raise EPIPE. And let's make it possible to skip over that
error.